float16 embeddings

+++ GraphRec-WWW19/Social_Encoders.py
@@ -26,7 +26,7 @@ class Social_Encoder(nn.Module):
             to_neighs.append(neighs)
         neigh_feats = self.aggregator.forward(nodes, to_neighs)  # user-user network
 
-        self_feats = self.features(torch.LongTensor(nodes.cpu().numpy())).to(self.device, dtype=torch.float16).float()
+        self_feats = self.features(torch.LongTensor(nodes.cpu().numpy())).to(self.device, dtype=torch.float16)
         self_feats = self_feats.t()
         
         # self-connection could be considered.
